requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
requests>=2.31.0
pyyaml>=6.0.0
rich>=13.0.0
